Update unit classification table on each app launch
The data can change, whether due to updates to the unit classification files
themselves, or because the app locale changed and we need to reload the correct
localised strings into the database.

As a quick fix, we therefore set the unit classifications to reload from the raw
.json files on each app launch, instead of only on the first app launch ever.

This new code in fact runs every time the activity gets recreated (unless the
recreation happens just while the previous task is still running), which use-
fully also handles locale changes while the app is already running (because a
locale change will destroy and recreate the activity, too).

In a subsequent commit, we can try to be smarter and only do the actual database
import when anything has actually changed.