Dump unit fingerprints via menu item
For more easily raising the target API version past 23, I want to get rid of the
WRITE_EXTERNAL_STORAGE permission, so instead, dumping the unit fingerprints can
now be triggered via a menu item and on API 19+, we dump the file into the
package-private external storage directory now, which still isn't too difficult
to access via USB or on a rooted phone.

On earlier APIs, we just dump the file into the internal files dir, because
that's the only place that doesn't require any permissions. On the off chance
that someone wants to retrieve the fingerprints from such an old device, it's a
bit more of a hassle, but with either adb or root access it should still work
one way or the other.