Bump target SDK level to 14
SDK 14 enables hardware acceleration by default, and its seems that 9-patch-
scaling is somehow handled differently between hardware and software rendering.

For the backspace button, we want the software rendering behaviour, which scales
up the image to fill the button size (maybe because the 9-patch borders them-
selves are all empty, so the centre content is scaled instead? Whereas hardware
rendering definitely seems to only scale the in this case empty borders and
leaves the image content in its original size).