Run InitialisationTask only after onPostCreate
If we already start in onCreate and start showing the loading dialog, then sub-
sequently Android will clobber the ManagedDialogs state during onRestore-
InstanceState and the call to dismiss the loading dialog again afterwards will
fail, because it is targetting the wrong dialog instance.

Therefore, we have to wait until onPostCreate before being able to launch the
task. (After super.onRestoreInstanceState() is the very earliest point that
would be safe, but onRestoreInstanceState() doesn't necessarily get called on
every activity launch. While we *are* using a flag to only kick of the
initialisation if we've actually gone through onCreate(), it's not worth
bothering just to kick it of minimally earlier.