Manually invalidate "All units" list after units classification reload
The framework automatically recreates all dialogs when the locale changes, how-
ever the classification strings in the database are only updated via an async
task that only gets to run afterwards.

Hence we manually need to track the fact that the unit classifications have
been reloaded and that we therefore need to reload the unit strings from the
database the next time the dialog is shown. (We could call removeDialog() to
force-invalidate it, but that would also close it if it happens to be open at
the time the database reload task completes.)