Upgrade units-in-java to 1.89.J01 - DONTBUILD
The copy of units-in-java in the source tree includes some additional changes
beyond the simple bulk edits in the upgrade makefile. We're going to reapply
those (as well as fix any other additional breakage that might have happened due
to the upgrade) in the next changeset.

String.empty() was only added in API level 9, so we bulk-edit it back to
String.length() == 0.

We unexclude the grammar file again, because further upstream updates are
unlikely and we might have to modify it at some point if we touch the units file
file parsing. Also it looks like the file ending up in the APK might have only
happened with the old build tools, but not our current build.