Re-run the unit usage task on version upgrades, too
The units auto-complete is powered by the unit usage database, so whenever the
underlying units data file might have changed, we need to re-run the task.
The units data can change both after an app update, if we've updated the data
file itself, or due to a locale change, because the unit definitions get
dynamically filtered depending on the current locale.

To also catch any units that might have been deleted in an update or locale
change, we create a temporary copy of the old usage data, clear out the database
table, repopulate based on the current units definitions and then copy any
existing usage values over.