Track float accuracy in unit list conversion and round off too small remainder
Example case was converting "1 guinea" to "poundsterling; shilling; pence",
which apart from the expected answer of 1 pound + 1 shilling also ended up with
a remainder of 1.7742632E-15 pence.

Therefore we attempt tracking the floating point accuracy through the scaling as
the unit conversion proceeds, and break off if the remainder drops below the
floating point accuracy.

This change also has the beneficial side-effect of aborting the calculation loop
once the remainder reaches zero, even if the actual performance benefit will be
minimal unless somebody enters a truly gigantic list of units to convert to.