Get rid of strings_calc
It seems those strings were originally imported from the AOSP calculator.
However except for recently repurposing the 'del' string as the accessible
description for the backspace button, we weren't actually using any of the
advanced strings, and most of the button texts weren't actually translateable
anyway.

Android Studio's translation editor also doesn't quite like string definitions
spread out across multiple files, so we just move the English button texts into
the main strings file and otherwise get rid of those files.

The 'del' string also wasn't ideal for the accessible description anyway,
because it was in all uppercase, and in German it was actually translated as
'C' (as in the corresponding calculator key on physical calculators).