Use moderately old support library
Preparation for switch to CursorLoaders.

Attempting to keep the minSdk version as low as possible, we need to attempt
using v25.4.0 of the old support library. It still means raising the minSdk
version from 4 to 9, but I suppose that's still old enough, and the oldest
emulator I can get to run is API10, anyway. (If somebody still seriously uses
Units on such an old phone, please get in touch.)